requirements.txt
aiofiles==24.1.0
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
bcrypt==4.1.3
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastecdsa==2.3.2
filelock==3.15.4
flower==2.0.1
fonttools==4.53.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
numpy==2.0.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
parsel==1.9.1
passlib==1.7.4
pillow==10.3.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
Protego==0.3.1
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytest==8.2.2
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
rabbitmq==0.2.0
redis==5.0.6
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
seaborn==0.13.2
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
tldextract==5.1.2
tomli==2.0.1
tornado==6.4.1
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
w3lib==2.2.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
zipp==3.19.2
zope.interface==6.4.post2
